iter4: fix tests
