+++ cmd/server/flags.go
@@ -1,0 +1,10 @@
+package main
+
+import "flag"
+
+var flagRunAddr string
+
+func parseFlags() {
+	flag.StringVar(&flagRunAddr, "a", "localhost:8080", "address and port to run server")
+	flag.Parse()
+}
